clean up

+++ train_triton.py
@@ -58,7 +58,7 @@ class TextSamplerDataset(Dataset):
 
     def __getitem__(self, index):
         rand_start = torch.randint(0, self.data.size(0) - self.seq_len, (1,))
-        full_seq = self.data[rand_start: rand_start + self.seq_len].long().type(torch.float16)
+        full_seq = self.data[rand_start: rand_start + self.seq_len].long()
         return full_seq.cuda()
 
     def __len__(self):
